+++ templates/navigation/index.php
@@ -1,5 +1,5 @@
-<ul>
-	<li><a href="prayertime"><b><?php echo $l->t('Prayer Time'); ?></b></a></li>
-        <li><a href="adjustments"><b><?php echo $l->t('Adjustments'); ?></b></a></li>
-        <li><a href="settings"><b><?php echo $l->t('Settings'); ?></b></a></li>
+<ul style="padding: calc(var(--default-grid-baseline)* 2);">
+	<li><a href="prayertime" class="app-navigation-entry"><b><?php echo $l->t('Prayer Time'); ?></b></a></li>
+        <li><a href="adjustments" class="app-navigation-entry"><b><?php echo $l->t('Adjustments'); ?></b></a></li>
+        <li><a href="settings" class="app-navigation-entry"><b><?php echo $l->t('Settings'); ?></b></a></li>
 </ul>